% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{log_message}
\alias{log_message}
\title{A simple logging function}
\usage{
log_message(message, mode = "INFO")
}
\arguments{
\item{message}{(string) The message to be logged}

\item{mode}{(string) type of log message (info / debug)}
}
\value{
none
}
\description{
A simple logging function
}
