% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minify_code.R
\name{minify_code}
\alias{minify_code}
\title{Capture and minify code context}
\usage{
minify_code(currentLine)
}
\arguments{
\item{currentLine}{(integer) The currently selected line}
}
\value{
(string) The code context, from line 0 to currentLine, minified
}
\description{
Captures and minifies the code context (string) in order to save on token budget.
The code context is all code in the editor above 'currentLine', up to line 0.
Currently minification is rudimentary to ensure best results on inference.
}
