% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_query.R
\name{replace_query}
\alias{replace_query}
\title{Replace query with completion}
\usage{
replace_query()
}
\value{
none
}
\description{
This function captures the selected text as a query, along with code context
between line 0 and the selected line. Then a prompt is generated, and the
selected query is replaced with the result from the model.
}
